detect serialized attribute changes
