+++ activemodel/lib/active_model/dirty.rb
@@ -159,7 +159,7 @@ module ActiveModel
 
     # Handle <tt>*_was</tt> for +method_missing+.
     def attribute_was(attr) # :nodoc:
-      attribute_changed?(attr) ? changed_attributes[attr] : __send__(attr)
+      changed_attributes.include?(attr) ? changed_attributes[attr] : __send__(attr)
     end
 
     private
@@ -178,7 +178,7 @@ module ActiveModel
 
       # Handle <tt>*_change</tt> for +method_missing+.
       def attribute_change(attr)
-        [changed_attributes[attr], __send__(attr)] if attribute_changed?(attr)
+        [changed_attributes[attr], __send__(attr)] if changed_attributes.include?(attr)
       end
 
       # Handle <tt>*_will_change!</tt> for +method_missing+.
